requirements.txt
pytest==7.4.0
selenium==4.12.0
pytest-xdist==3.3.1
python-dotenv==1.0.0
